requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
manim
supabase==2.0.2